requirements.txt
google-genai
tqdm
python-dotenv